+++ include/dungeon/dungeon.h
@@ -22,6 +22,7 @@
 #include <stdint.h>
 #include <GL/glew.h>
 
+struct page;
 struct dgn;
 struct dtx;
 struct tes;
@@ -61,6 +62,10 @@ void dungeon_context_free(struct dungeon_context *ctx);
 bool dungeon_load(struct dungeon_context *ctx, int num);
 void dungeon_set_camera(int surface, float x, float y, float z, float angle, float angle_p);
 void dungeon_render(struct dungeon_context *ctx);
+void dungeon_set_walked(int surface, int x, int y, int z, int flag);
+int dungeon_calc_conquer(int surface);
+bool dungeon_load_walk_data(int surface, int map, struct page **page);
+bool dungeon_save_walk_data(int surface, int map, struct page **page);
 
 struct dungeon_context *dungeon_get_context(int surface);
 void dungeon_update(void);
