+++ src/dungeon/dgn.c
@@ -42,7 +42,7 @@ struct dgn *dgn_parse(uint8_t *data, size_t size)
 	dgn->size_z = buffer_read_int32(&r);
 	buffer_skip(&r, 40);
 
-	int nr_cells = dgn->size_x * dgn->size_y * dgn->size_z;
+	int nr_cells = dgn_nr_cells(dgn);
 	dgn->cells = xcalloc(nr_cells, sizeof(struct dgn_cell));
 	uint32_t x = 0, y = 0, z = 0;
 	for (int i = 0; i < nr_cells; i++) {
@@ -146,7 +146,7 @@ struct dgn *dgn_parse(uint8_t *data, size_t size)
 
 void dgn_free(struct dgn *dgn)
 {
-	int nr_cells = dgn->size_x * dgn->size_y * dgn->size_z;
+	int nr_cells = dgn_nr_cells(dgn);
 
 	for (int i = 0; i < nr_cells; i++) {
 		if (dgn->cells[i].visible_cells)
@@ -173,6 +173,21 @@ int dgn_cell_index(struct dgn *dgn, uint32_t x, uint32_t y, uint32_t z)
 struct dgn_cell *dgn_cell_at(struct dgn *dgn, uint32_t x, uint32_t y, uint32_t z)
 {
 	return &dgn->cells[dgn_cell_index(dgn, x, y, z)];
+}
+
+int dgn_calc_conquer(struct dgn *dgn)
+{
+	int nr_cells = dgn_nr_cells(dgn);
+	int enterable = 0;
+	int walked = 0;
+	for (int i = 0; i < nr_cells; i++) {
+		if (!dgn->cells[i].enterable)
+			continue;
+		enterable++;
+		if (dgn->cells[i].walked)
+			walked++;
+	}
+	return walked * 100 / enterable;
 }
 
 struct pvs_cell {
