+++ src/hll/DrawDungeon.c
@@ -360,12 +360,12 @@ HLL_LIBRARY(DrawDungeon,
 			HLL_EXPORT(SetDrawMapRadar, dungeon_map_set_radar),
 			HLL_EXPORT(SetDrawHalfFlag, DrawDungeon_SetDrawHalfFlag),
 			HLL_EXPORT(GetDrawHalfFlag, DrawDungeon_GetDrawHalfFlag),
-			HLL_EXPORT(SetWalked, dungeon_map_set_walked),
+			HLL_EXPORT(SetWalked, dungeon_set_walked),
 			HLL_TODO_EXPORT(SetWalkedAll, DrawDungeon_SetWalkedAll),
 			HLL_TODO_EXPORT(CalcNumofFloor, DrawDungeon_CalcNumofFloor),
 			HLL_TODO_EXPORT(CalcNumofWalk, DrawDungeon_CalcNumofWalk),
 			HLL_TODO_EXPORT(CalcNumofWalk2, DrawDungeon_CalcNumofWalk2),
-			HLL_EXPORT(CalcConquer, dungeon_map_calc_conquer),
+			HLL_EXPORT(CalcConquer, dungeon_calc_conquer),
 			HLL_EXPORT(SetInterlaceMode, DrawDungeon_SetInterlaceMode),
 			HLL_EXPORT(SetDirect3DMode, DrawDungeon_SetDirect3DMode),
 			HLL_EXPORT(SaveDrawSettingFlag, DrawDungeon_SaveDrawSettingFlag),
@@ -378,6 +378,6 @@ HLL_LIBRARY(DrawDungeon,
 			HLL_EXPORT(GetTexSound, DrawDungeon_GetTexSound),
 			HLL_EXPORT(StopTimer, DrawDungeon_StopTimer),
 			HLL_EXPORT(RestartTimer, DrawDungeon_RestartTimer),
-			HLL_EXPORT(LoadWalkData, dungeon_map_load_walk_data),
-			HLL_EXPORT(SaveWalkData, dungeon_map_save_walk_data)
+			HLL_EXPORT(LoadWalkData, dungeon_load_walk_data),
+			HLL_EXPORT(SaveWalkData, dungeon_save_walk_data)
 			);
